Refactor DataSource test cases
Call testsuite functions to refactor test cases
