+++ test/helpers/config.json
@@ -1,0 +1,16 @@
+ {
+  "DefaultModelsMeta": {
+     "sources": [
+       "loopback/common/models",
+       "loopback/server/models",
+       "../common/models",
+       "./models"
+     ],
+     "mixins": [
+       "loopback/common/mixins",
+       "loopback/server/mixins",
+       "../common/mixins",
+       "./mixins"
+     ]
+   }
+ }
